Use ModuleHandle for module management in Executor
We used to use void pointers to identify loaded GPU binaries. This change makes it use the `ModuleHandle` type. It avoids a bunch of `reinterpret_cast` and makes it also more clear what the void pointer represents.

In addition this is replacing the out parameters in a bunch of related functions by `absl::StatusOr<...>` return types.

PiperOrigin-RevId: 685849586